+++ R/cran.R
@@ -1,2 +1,0 @@
-
-dep_fields <- c("Depends", "Imports", "Suggests", "Enhances", "LinkingTo")